fix: moved tabs component in tabs folder
