+++ src/components/Tabs/Tabs.story.vue
